(Compliance) Refactor: rename package warehouse to store
