+++ src/test/java/store/ProductListTest.java
@@ -1,4 +1,4 @@
-package warehouse;
+package store;
 
 import org.junit.Test;
