+++ src/test/java/store/ProductTest.java
@@ -1,4 +1,4 @@
-package warehouse;
+package store;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -10,6 +10,7 @@ import java.util.ArrayList;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+
 
 public class ProductTest {
