+++ src/test/java/store/ProviderTest.java
@@ -1,8 +1,9 @@
-package warehouse;
+package store;
 
 import static org.junit.Assert.*;
 
 import org.junit.Test;
+
 
 public class ProviderTest {
