+++ src/test/java/store/StockableProductTest.java
@@ -1,4 +1,4 @@
-package warehouse;
+package store;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
+
 
 class StockableProductTest {
