+++ src/test/java/store/WarehouseElementTest.java
@@ -1,4 +1,4 @@
-package warehouse;
+package store;
 
 import java.util.ArrayList;
 import java.util.stream.Stream;
